// src/types/processed.ts
export type ProcessedDbType = 
  | "MagnetAxiom"
  | "CellebritePA"
  | "XWays"
  | "Autopsy"
  | "EnCase"
  | "FTK"
  | "GenericSqlite"
  | "Unknown";